\name{iTable-package}
\alias{iTable-package}
\alias{iTable}
\docType{package}
\title{
\packageTitle{iTable}
}
\description{
\packageDescription{iTable}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{iTable}
\packageIndices{iTable}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{iTable}

Maintainer: \packageMaintainer{iTable}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
